Add event handling for CLOSE & PRICE
The tests introduced int his chapter work towards the feature of joining an auction and losing without bidding.
This adds the AuctionMessageTranslatorTest, which verifies that an auction is closed on a. CLOSE event and the current price gets updated on a PRICE event.
The end-to-end test is broken due to no implementation for the new currentPrice() method. This is mocked in the test and the functionality isn't relevant to the event handling, so no implementation is needed yet.